+++ android/app/src/main/java/com/xianxia/sect/core/engine/service/EventService.kt
@@ -145,7 +145,7 @@ class EventService @Inject constructor(
                 "equipment" -> {
                     val equipment = EquipmentDatabase.generateRandom(rarity, rarity)
                     val template = EquipmentDatabase.getTemplateByName(equipment.name)
-                    val basePrice = ((template?.price ?: GameConfig.Rarity.get(rarity).basePrice) * GameConfig.Rarity.PRICE_MULTIPLIER).roundToInt()
+                    val basePrice = ((template?.price ?: GameConfig.Rarity.get(rarity).basePrice) * GameConfig.Rarity.PRICE_MULTIPLIER).roundToInt().toLong()
                     MerchantItem(
                         id = UUID.randomUUID().toString(),
                         name = equipment.name,
@@ -161,7 +161,7 @@ class EventService @Inject constructor(
                 "manual" -> {
                     val manual = ManualDatabase.generateRandom(rarity, rarity)
                     val template = ManualDatabase.getByName(manual.name)
-                    val basePrice = ((template?.price ?: GameConfig.Rarity.get(rarity).basePrice) * GameConfig.Rarity.PRICE_MULTIPLIER).roundToInt()
+                    val basePrice = ((template?.price ?: GameConfig.Rarity.get(rarity).basePrice) * GameConfig.Rarity.PRICE_MULTIPLIER).roundToInt().toLong()
                     MerchantItem(
                         id = UUID.randomUUID().toString(),
                         name = manual.name,
@@ -179,7 +179,7 @@ class EventService @Inject constructor(
                     if (pillTemplates.isEmpty()) continue
                     val template = pillTemplates.random(random)
                     val pill = ItemDatabase.createPillFromTemplate(template)
-                    val basePrice = (template.price * GameConfig.Rarity.PRICE_MULTIPLIER).roundToInt()
+                    val basePrice = (template.price * GameConfig.Rarity.PRICE_MULTIPLIER).roundToInt().toLong()
                     MerchantItem(
                         id = UUID.randomUUID().toString(),
                         name = pill.name,
@@ -197,7 +197,7 @@ class EventService @Inject constructor(
                     val materials = BeastMaterialDatabase.getMaterialsByRarity(rarity)
                     if (materials.isEmpty()) continue
                     val material = materials.random(random)
-                    val basePrice = (material.price * GameConfig.Rarity.PRICE_MULTIPLIER).roundToInt()
+                    val basePrice = (material.price * GameConfig.Rarity.PRICE_MULTIPLIER).roundToInt().toLong()
                     MerchantItem(
                         id = UUID.randomUUID().toString(),
                         name = material.name,
@@ -214,7 +214,7 @@ class EventService @Inject constructor(
                     val herbs = HerbDatabase.getByRarity(rarity)
                     if (herbs.isEmpty()) continue
                     val herb = herbs.random(random)
-                    val basePrice = (herb.price * GameConfig.Rarity.PRICE_MULTIPLIER).roundToInt()
+                    val basePrice = (herb.price * GameConfig.Rarity.PRICE_MULTIPLIER).roundToInt().toLong()
                     MerchantItem(
                         id = UUID.randomUUID().toString(),
                         name = herb.name,
@@ -231,7 +231,7 @@ class EventService @Inject constructor(
                     val seeds = HerbDatabase.getSeedsByRarity(rarity)
                     if (seeds.isEmpty()) continue
                     val seed = seeds.random(random)
-                    val basePrice = (seed.price * GameConfig.Rarity.PRICE_MULTIPLIER).roundToInt()
+                    val basePrice = (seed.price * GameConfig.Rarity.PRICE_MULTIPLIER).roundToInt().toLong()
                     MerchantItem(
                         id = UUID.randomUUID().toString(),
                         name = seed.name,
@@ -295,7 +295,7 @@ class EventService @Inject constructor(
         val sect: WorldSect,
         val item: MerchantItem,
         val actualQuantity: Int,
-        val totalPrice: Int,
+        val totalPrice: Long,
         val updatedSectDetails: Map<String, SectDetail>
     )
 
@@ -313,7 +313,7 @@ class EventService @Inject constructor(
 
         val actualQuantity = minOf(quantity, item.quantity)
         val priceMultiplier = calculatePriceMultiplier(data, sectId)
-        val totalPrice = (item.price * priceMultiplier).toInt() * actualQuantity
+        val totalPrice = (item.price * priceMultiplier).toLong() * actualQuantity
 
         if (data.spiritStones < totalPrice) {
             addGameEvent("灵石不足，无法购买${item.name}", EventType.WARNING)
